bugfixes+ introduction of verbose level
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder@1723 1f5c12ca-751b-0410-a591-d2e778427230

+++ trunk/mert/Point.cpp
@@ -15,7 +15,7 @@ unsigned Point::ncall=0;
 
 void Point::Randomize(const parameters_t& min,const parameters_t& max){
   for (int i=0; i<size(); i++)
-    operator[](i)= min[i] + (float)random()/RAND_MAX * (max[i]-min[i]);
+    operator[](i)= min[i] + (float)random()/(float)RAND_MAX * (float)(max[i]-min[i]);
 }
 
 void Point::Normalize(){
